requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
chromadb
openai
python-dotenv
pyyaml
pandas
httpx
pytest
groq
